# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trak fitness coaching platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /programs:
    get:
      operationId: listPrograms
      tags: [workouts]
      summary: List all workout programs
      responses:
        "200":
          description: Programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [workouts]
      summary: Create a workout program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /programs/{programId}:
    get:
      operationId: getProgram
      tags: [workouts]
      summary: Get a program with its days and exercises
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDetail"
    patch:
      operationId: updateProgram
      tags: [workouts]
      summary: Update a program
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    delete:
      operationId: deleteProgram
      tags: [workouts]
      summary: Delete a program
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /programs/{programId}/phases:
    post:
      operationId: createProgramPhase
      tags: [workouts]
      summary: Add a phase to a program
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramPhaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramPhase"
  /programs/{programId}/phases/{phaseId}:
    patch:
      operationId: updateProgramPhase
      tags: [workouts]
      summary: Update a program phase
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
        - name: phaseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramPhaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramPhase"
    delete:
      operationId: deleteProgramPhase
      tags: [workouts]
      summary: Delete a program phase
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
        - name: phaseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /programs/{programId}/days:
    post:
      operationId: createProgramDay
      tags: [workouts]
      summary: Add a day to a program
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramDayInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDay"
  /programs/{programId}/days/{dayId}:
    patch:
      operationId: updateProgramDay
      tags: [workouts]
      summary: Update a program day
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramDayUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDay"
    delete:
      operationId: deleteProgramDay
      tags: [workouts]
      summary: Delete a program day
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /programs/{programId}/days/{dayId}/exercises:
    post:
      operationId: addExerciseToDay
      tags: [workouts]
      summary: Add an exercise to a program day
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramExerciseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramExercise"
  /programs/{programId}/days/{dayId}/exercises/{peId}:
    patch:
      operationId: updateProgramExercise
      tags: [workouts]
      summary: Update a program day exercise
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
        - name: peId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramExerciseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramExercise"
    delete:
      operationId: deleteProgramExercise
      tags: [workouts]
      summary: Delete a program day exercise
      parameters:
        - name: programId
          in: path
          required: true
          schema:
            type: integer
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
        - name: peId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Program Assignments ───────────────────────────────────────────────────
  /clients/{clientId}/program-assignment:
    get:
      operationId: getClientProgramAssignment
      tags: [workouts]
      summary: Get the active program assigned to a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Active assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramAssignment"
        "404":
          description: No active program
    post:
      operationId: assignProgram
      tags: [workouts]
      summary: Assign a program to a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramAssignmentInput"
      responses:
        "201":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramAssignment"

  # ── Workout Logs ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Program:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        durationWeeks:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ProgramInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        durationWeeks:
          type: integer
    ProgramUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        durationWeeks:
          type: ["integer", "null"]
    ProgramPhase:
      type: object
      required: [id, programId, name, order, durationWeeks]
      properties:
        id:
          type: integer
        programId:
          type: integer
        name:
          type: string
        order:
          type: integer
        durationWeeks:
          type: integer
        daysPerWeek:
          type: integer
          nullable: true
    ProgramPhaseInput:
      type: object
      required: [name, durationWeeks]
      properties:
        name:
          type: string
          minLength: 1
        durationWeeks:
          type: integer
          minimum: 1
        daysPerWeek:
          type: integer
          minimum: 1
          maximum: 7
          nullable: true
        order:
          type: integer
    ProgramPhaseUpdate:
      type: object
      properties:
        name:
          type: string
        durationWeeks:
          type: integer
        daysPerWeek:
          type: integer
          minimum: 1
          maximum: 7
          nullable: true
        order:
          type: integer
    ProgramPhaseDetail:
      type: object
      required: [id, name, order, durationWeeks, days]
      properties:
        id:
          type: integer
        name:
          type: string
        order:
          type: integer
        durationWeeks:
          type: integer
        daysPerWeek:
          type: integer
          nullable: true
        days:
          type: array
          items:
            $ref: "#/components/schemas/ProgramDayDetail"
    ProgramDetail:
      type: object
      required: [id, name, days, phases]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        durationWeeks:
          type: ["integer", "null"]
        phases:
          type: array
          items:
            $ref: "#/components/schemas/ProgramPhaseDetail"
        days:
          type: array
          items:
            $ref: "#/components/schemas/ProgramDayDetail"
        createdAt:
          type: string
          format: date-time
    ProgramDay:
      type: object
      required: [id, programId, dayNumber, name]
      properties:
        id:
          type: integer
        programId:
          type: integer
        dayNumber:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
        phaseId:
          type: ["integer", "null"]
    ProgramDayInput:
      type: object
      required: [dayNumber, name]
      properties:
        dayNumber:
          type: integer
        name:
          type: string
          minLength: 1
        notes:
          type: string
        phaseId:
          type: integer
    ProgramDayUpdate:
      type: object
      properties:
        dayNumber:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
        phaseId:
          type: ["integer", "null"]
    ProgramDayDetail:
      type: object
      required: [id, dayNumber, name, exercises]
      properties:
        id:
          type: integer
        dayNumber:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
        phaseId:
          type: ["integer", "null"]
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/ProgramExerciseDetail"
    ProgramExercise:
      type: object
      required: [id, dayId, exerciseId, sets, reps, order]
      properties:
        id:
          type: integer
        dayId:
          type: integer
        exerciseId:
          type: integer
        sets:
          type: integer
        reps:
          type: string
        order:
          type: integer
        weight:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        restSeconds:
          type: ["integer", "null"]
    ProgramExerciseInput:
      type: object
      required: [exerciseId, sets, reps, order]
      properties:
        exerciseId:
          type: integer
        sets:
          type: integer
        reps:
          type: string
          minLength: 1
        order:
          type: integer
        weight:
          type: string
        notes:
          type: string
        restSeconds:
          type: integer
    ProgramExerciseUpdate:
      type: object
      properties:
        sets:
          type: integer
        reps:
          type: string
        order:
          type: integer
        weight:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        restSeconds:
          type: ["integer", "null"]
    ProgramExerciseDetail:
      type: object
      required: [id, exerciseId, exerciseName, muscleGroup, sets, reps, order]
      properties:
        id:
          type: integer
        exerciseId:
          type: integer
        exerciseName:
          type: string
        muscleGroup:
          type: string
        sets:
          type: integer
        reps:
          type: string
        order:
          type: integer
        weight:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        restSeconds:
          type: ["integer", "null"]

    # ── Program Assignment ──────────────────────────────────────────────────
    ProgramAssignment:
      type: object
      required: [id, clientId, programId, programName, startDate]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        programId:
          type: integer
        programName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
    ProgramAssignmentInput:
      type: object
      required: [programId, startDate]
      properties:
        programId:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    # ── Workout Logs ────────────────────────────────────────────────────────
...